This is the emulator for the Ferranti
This is a "one file", standalone emulator for the Ferranti Mercury. There is a lot of work to be done yet, in order to make the "Input Routine" fully functional. You see, the "Annotated Input Routine" was written (handwritten) in a kind of "relative addressing mode", so to make the binaries functional one must find a suitable place in the "drum memory", then adjust all addresses accordingly. There's some kind of code too, you must become a "human assembler" to complete this primitive "operating system". But don't get me wrong: this code has months of hard work worth, so it is a good "starting point". Have fun..! Jullián